Make file offset non-optional
For files, you have to pass -1 for the current position, but that's an
error with sockets. For sockets, passing 0 reads the next thing, but for
files that's absolute position 0. So, make the user say which one they
want, as we can't provide a sensible default.

Also, rename `offset` to `file_offset` in some places for consistency.