Don't allow freeing the ring while it's in use
Add the ring to a global set on creation and remove it on exit.
This prevents GC'ing buffers that we have passed to the kernel.
Attempting to call `exit` while the ring is in use is now an error.

Also, check for invalid cancel requests before allocating a slot.

Note: uring now requires OCaml >= 4.12 for the Atomic module.