Implement Uring.read and Uring.write
These are simple wrappers for read(2) and write(2) that use a Cstruct.t.
This avoids the extra allocations done by readv(2) and writev(2) as we currently
need to do one malloc per readv(2)/writev(2).

It's also more natural for most applications that only have a single buffer to
read/write.