Introduce a Sketch buffer per Uring.
The main motivation of this change is to avoid having one malloc per packet in
readv(2), writev(2) & friends.

Before this, everytime we needed an iovec structure we would malloc it outside
the heap, then fill it up and set up the GC to collect it.

Often we only need those structures in the window between the
`Uring.submit_writev` and `Uring.submit`, once the submit is done, the kernel
has already copied the arguments and we don't need them anymore. In the old code
the iovec would end up being held until atleast the CQEvent, this is not a big
deal, but mallocing per syscall kinda is.

The sketch buffer is a buffer that starts with length zero and we grow it every
time we need more space and we never shrink it back, the assumption is that if
the application needed that much space at one point, it will require it again in
the future.

The memory allocated is freed on `Uring.submit` as long as we can guarantee all
submits are not pending, since we allocate a new buffer when we need to grow it,
we also need to release old buffers that were used in-between.

The allocations are aligned to the machine's word size, this is more of a safe
guard to guarantee we could run on a strict aligned machine, and also to gain
some cycles as unaligned access are still slower on x86 and and arm.

The `main.md` tests make sure the buffer can grow and we don't crash.